pyproject.toml
[tool.poetry]
name = "typeshift"
version = "0.1.0"
description = ""
authors = ["Joel Grus <joelgrus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
bitarray-hardbyte = "^1.6.2"
streamlit = "^0.77.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
